Completed NotesTable tests
