+++ src/__tests__/incidents/view/ViewIncidentDetails.test.tsx
@@ -1,4 +1,3 @@
-import { Button } from '@hospitalrun/components'
 import { mount, ReactWrapper } from 'enzyme'
 import { createMemoryHistory } from 'history'
 import React from 'react'
